FEAT(bootcamp): completed bootcamp

+++ server/src/routes/memories.ts
@@ -1,52 +1,74 @@
-import { FastifyInstance } from 'fastify';
-import { z } from 'zod';
-import { prisma } from '../lib/prisma';
+import { FastifyInstance } from 'fastify'
+import { z } from 'zod'
+import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
       orderBy: {
         createdAt: 'asc',
       },
-    });
+      where: {
+        userId: request.user.sub,
+      },
+    })
 
     return memories.map((memory) => {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
         excerpt: memory.content.substring(0, 115).concat('...'),
-      };
-    });
-  });
+        createdAt: memory.createdAt,
+      }
+    })
+  })
 
-  app.get('/memories/:id', async (request) => {
-    const paramsSchema = z.object({ id: z.string().uuid() });
+  app.get('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({ id: z.string().uuid() })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
-    });
+    })
 
-    return memory;
-  });
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
-  app.put('/memories/:id', async (request) => {
-    const paramsSchema = z.object({ id: z.string().uuid() });
+    return memory
+  })
 
-    const { id } = paramsSchema.parse(request.params);
+  app.put('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({ id: z.string().uuid() })
+
+    const { id } = paramsSchema.parse(request.params)
 
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean(),
-    });
+    })
 
-    const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
+    const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -54,47 +76,56 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '88a078f9-ad5c-4385-af87-c8f1b582a34a',
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
   app.post('/memories', async (request) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean(),
-    });
+    })
 
-    const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
+    const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: '88a078f9-ad5c-4385-af87-c8f1b582a34a',
+        userId: request.user.sub,
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
-  app.delete('/memories/:id', async (request) => {
-    const paramsSchema = z.object({ id: z.string().uuid() });
-    const { id } = paramsSchema.parse(request.params);
+  app.delete('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({ id: z.string().uuid() })
+    const { id } = paramsSchema.parse(request.params)
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
         id,
       },
-    });
+    })
 
     return {
       status: 200,
       message: `Memory with id ${id} is deleted`,
-    };
-  });
+    }
+  })
 }
